Rob mentioned an instance where a query sequence with highly biased composition (a flocculin protein?  heavy on S and T) had caused a timeout on the webserver. Running phmmer on that sequence against pfamseq I noticed that MSV passed 1086999 sequences (~11.5% of the db) and the bias filter didn't remove any of them. Running hmmsearch produced wildly different, with the bias filter acting on a large chunk of msv-passing sequences, leaving only 2% of sequence. I have a hard time imagining that Sean meant for the bias filter to be disabled (it isn't documented as such).
The culprit: it appears that phmmer never set the compo values, upon which the
bias filter depends. This commit just adds that step to p7_SingleBuilder(), via 
a call to p7_hmm_SetComposition() and subsequent copying of compo from the hmm 
to the model (om).

The result is that a phmmer-to-pfamseq search time on this query is reduced 
from 11,155 seconds to 846, and that 15 marginal hits are lost (212 vs 227, 
with most-significant loss having e-value of 0.031).

Data files are available in ~wheelert/notebook/2011/0628_hmmerweb_timeout

git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3589 db3e70e4-12e1-0310-b23f-81a6e3d09ba7