revised TJW's fix for p7_SingleBuilder() not setting compo[]. No need to manually copy it into the P7_OPROFILE; oprofile construction takes care of it, if compo has been set when it should be (by moving the p7_hmm_SetComposition() call prior to oprofile construction).
Profmark benchmark suggests that setting compo[] degrades phmmer
performance -- as expected, because this is the tradeoff of using the
bias filter. But the degradation is minor, and is more than
compensated by the recent (14 June) change that removed Yu/Altschul
calculations and improved performance. Overall, phmmer performance is
still improved relative to 3.0 release, so I'm not much concerned
about the performance loss of using the bias filter in phmmer.

fixed two compiler warnings: typo in nhmmer.c options structure,
unused counter in p7_tophits.c



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3593 db3e70e4-12e1-0310-b23f-81a6e3d09ba7