Corrected p7_oprofile_ReadInfoMSV() so that it is now in sync with p7_oprofile_Write(). Previously, <om->max_length> and SSV match scores (<om->sbv>) were not being accounted for, presumably because this function wasn't updated when those variables were added to the oprofile object and/or output.
p7_oprofile_ReadInfoMSV() seems to only be used by MPI hmmscan in
hmmer to determine file offsets in hmmpress'ed MSV filter files (.h3f)
to send to workers. The workers are to start reading filter info at
the received offset. 

I came across this when writing cmpress and cmscan, the latter of
which uses this function to skip over any unwanted p7 filters that
exist in the cmpress'ed MSV filter (.i1f) file. Eventually it will be
used in MPI cmscan as it is in MPI hmmscan, as well.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3606 db3e70e4-12e1-0310-b23f-81a6e3d09ba7