update p7_alidisplay_EncodeAliPostProb with new APP translation (mapping the posterior prob of being aligned to the model to integers 0..9 plus '*').
Also fixed an off-by-one error that caused the APP line to be 
shifted two to the left.

Using these values to trim alignments in my dmark ghost benchmark, 
there's a nice sweet spot where we lose roughly 4 true-positive
bases for every 3 false-positive bases we throw out. The result is 
that it's pretty easy to reproduce blastn's overextension 
specificity while retaining much higher sensitivity.

Not expected to impact anything but nhmmer. 



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3615 db3e70e4-12e1-0310-b23f-81a6e3d09ba7