Altered the CFLAGS value in the configure line of the ppc-aix-xlc autobuild script so that makenhmmerdb and exactmatch can allocate enough memory to work.
It seems by default, and AIX process is only allocated 256MB, of 
which ~1/2 is given to heap - this causes malloc to return a null 
pointer for any malloc call that causes total used heap to exceed 
~128MB (even for a very simple hello_world/malloc program).

The workaround is to add a compile flag (-bmaxdata:0x80000000) to the
xlc_r call, which allows ~2GB of allocated memory.  That's still
surprisingly small, but I'm not able to find a documented way to 
increase it.


Also added timing code to makenhmmerdb.c



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3647 db3e70e4-12e1-0310-b23f-81a6e3d09ba7