exactmatch.c now handles the new FM-index-set format produced by makenhmmerdb.c. The change to makenhmmerdb was essentially about 2 things: (1) using of easel's sequence I/O, and (2) partitioning data into a collection of FM-indexes, to facilitate     parallelization.
The code is in place to (in my tests so far) correctly retrieve the sequence 
name/id and position for each FM match.

At this point, makenhmmerdb can convert a fasta-formatted file to a collection
of FM-indexes (stored in a single file), which exactmatch can then use to search
strings that are identical to a query sequence (or collection of such sequence).



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3675 db3e70e4-12e1-0310-b23f-81a6e3d09ba7