Implemented the Forward/Backward search approach on a bi-directional FM-index described in   Simpson (Bioinformatics 2010) and   Lam (IEEE ICBB 2009)
Notes in ~/notebook/2011/0906_fmindex_hmmer2/

Basically, this allows me now to search for both forward and reverse instances of a query string, and keep track of the matching range in the suffix-array corresponding to the input database text T while growing the search string (required for the nhmmer string-tree-traversal approach to be used). In practice, the forward search is ~40% slower than the reverse search, at a savings of something like 30% space.  I may end up dumping this implementaion in favor of a faster one that uses more space ... but will stick with this one for now.

Haven't had a chance to test the altivec implementation, so I won't be surprised if the nightly build croaks. Will address that over the weekend.





git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3686 db3e70e4-12e1-0310-b23f-81a6e3d09ba7