(1) Switched my editor (eclipse) to never use tabs, then removed tabs from fm-related files I've been working on (the ones commited here)
(2) Placed global FM variables into a struct that is then passed around to functions as needed. This matches HMMER's standard of avoiding global and static variables, and should reduce problems with upcoming parallelization. This required a number of changes, including some to the implementation-specific code.  imple_sse checks out fine, but I haven't yet tested dummy or vmx, which undoubtedly will suffer some small errors, to be fixed in the AM.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3689 db3e70e4-12e1-0310-b23f-81a6e3d09ba7