Fixed error in the way the final few characters of text and BWT were stored by makenhmmerdb. This was causing boundary-case errors in recovering positions of exact match hits.
Clarified memory allocation sizes.

Corrected calculation of hit positions in the new fowrard/reverse FM-index data 
structure.  Since an FM-index is often the concatenation of a number of database
sequence, this involved back-calculating the originating sequence and position 
within that sequence, using binary search on the start-positions of each 
sequence in the concatenation (this is faster than the alternative of storing in
a way that can be retrieved during reading from the sampled suffix array).

Hits for a single query are now sorted by target sequence, and position within
target.

Test case has been added to the testsuite: i20-fmindex-core.pl.  This doesn't
test nhmmer integration, just the functionality of makenhmmerdb and the core
FM-index functions that read from it.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3696 db3e70e4-12e1-0310-b23f-81a6e3d09ba7