Fixed error in tophits duplicate merge
Fixed error in which FM-index memory was being repeatedly re-allocated.

Fixed a couple of errors related to diagonal length-calculation that caused writes to out-of-bounds positions in an array, which itself cause surprising downstream errors.

Correctly handle positions for complementary hits.

Some hits that should be found are still not making it through the pipeline, so this is not yet ready for primetime.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3727 db3e70e4-12e1-0310-b23f-81a6e3d09ba7