FM-index code still in progress. This work primarily modifies the SIMD-MSV code to make it easier to allow SIMD- and FM-MSV methods to extract the same windows from a full-length DNA sequence, with the result being consistency in downstream envelopes and e-values.
To do this, the SSV code now expends the energy to find which diagonal
is responsible for reaching the MSV-passing score threshold, then extends
the diagonal in a way analagous to FM-SSV's approach, and uses the bounds 
of this extended diagonal when establishing the surrounding window.

Updated testsuite to agree with the slightly different pipeline outcomes
produced by this new approach.


This will temporarily break things:
* FM-index code is currently broken, as I bring it into agreement with
the SIMD-code. 
* The VMX and dummy code isn't yet updated. That means they'll both fail
'make check', simply because the fraction of sequence passing pipeline
stages disagrees with my new changes.  I expect this to be fixed in
a day or two.

This won't impact the testsuite, but I also still need to modify the MSV 
code to handle diagonal recovery.

Benefits of going the route of diagonal-recovery:
- with diagonals in hand, I can do bias-composition filtering on 
  individual diagonals (instead of windows around them), which I expect
  to help with a problem we're seeing in repeat-search.
  
- Looking forward, these may also prove useful for establishing bands 
  for banded-Viterbi, should we decide to go that route.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3749 db3e70e4-12e1-0310-b23f-81a6e3d09ba7