Numerous fixes to in progress MSV code changes, commited so I can copy to login-eddy for tests.
SIMD SSV diagonal finding had a serious over-extension error in backtracking,
now resolved

FM MSV migrating towards code that will allow it to better mirror the 
diagonals and windows produced by SIMD MSV. In addition to lots of minor
changes, there was an error in the way it chose bounds for diagonal recovery, 
causing very long diagonals (i.e. really good hits) to be lost. I've fixed it 
here, though the current implementation will be replaced in a couple days with
a method that will allow better agreement of the diagonals produced by SIMD
and FM MSVs.

This code has a workaround for an error caused by building an FM-index on
an overly long chunk of text (the code is grabbing more sequence at a time
than it's supposed to). An actual fix is forthcoming - this workaround is just
to allow testing of other code in progress.

NOTE:  SSV is working in both SSE and FM.  MSV is almost certainly broken 
(it's a rare case, and doesn't impact my repeat HMMs or rmark3).  Also, 
VMX and dummy implementations are currently broken. Those will be dealt 
with next.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3769 db3e70e4-12e1-0310-b23f-81a6e3d09ba7