Changed indentation in a large number of lines, replacing tabs with spaces.
The only change of substance is one that is a workaround for a
printing bug in gcc on OS X, related to printing the exp(v) for
a substantially negative v, after having called 
_MM_SET_FLUSH_ZERO_MODE(_MM_FLUSH_ZERO_ON); 

See ~wheelert/notebook/2012/0106_printf_underflow_bug for details.
The upshot is that, when v == -735, the line
    printf ("%.0f:  %9.2g\n", v, exp(v));
should print     
    6.2e-320
but once _MM_SET_FLUSH_ZERO_MODE has been called as above, it prints
       e-321
              
This reproduces on my laptop and Sean's desktop. For number of less 
magnitude (e.g. -734 or greater), this error isn't seen.


The workaround I've chosen to use is to avoid directly
printing exp(v), and instead compute a value
   P = v < -700 ? 0 : exp(v)
then print based on P.

This has the added advantage of making output on login-eddy
and OS X systems congruent (login eddy truncates to 0 much
earlier than OS X does, setting exp(v) to 0 at around -708 
instead of around -746 as on my laptop).





git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3802 db3e70e4-12e1-0310-b23f-81a6e3d09ba7