(1) added functionality in hmmbuild to build a model for single sequence inputs using substitution matrix and gap scores as is done in phmmer and the first iteration of jackhmmer.
This is expected to work for any single-sequence "alignment" provided
as input, including one in the middle of a concatenated list of
alignments as is possible in stockholm format files. Single- and
multi-threaded implementations are tested, mpi is implemented but
not tested (it's a pretty simple change, so I expect it to just work
... but I need to discuss mpi testing with Sean).

In this new method, the --single flag causes single-sequence
"alignments" to have models built as phmmer would build them, i.e.
not using the mixture dirichlet priors. Model construction
for multiple sequence alignment inputs will not be affected by
that flag.

(2) In doing this, I observed that the matrices produced by
p7_builder_(Load|Set)ScoreSystem() did not pass the p7_hmm_Validate()
test, as p(b|a) matrix values are normalized to sum to 1 for a
fixed a. I've now modified those functions to normalize scores 
before returning, so that validate now passes.

'make check' passes, as does 'make dev'. I expect this will probably
have some sort of impact on phmmer benchmark results, if these exist.
Do we have these? 



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3819 db3e70e4-12e1-0310-b23f-81a6e3d09ba7