Two modifications to the way nhmmer uses the domain definition machinery:
(1) In region_trace_ensemble(), 200 traces are sampled from the 
posterior probability matrix, and used in picking domain boundaries. 
I assume that works fine for relatively short (e.g protein) sequences,
but it seems like a potential problem in the case of dealing with a 
very large window with many true hits. Alu on a 1Mb sequence chunk 
is likely to find windows containing 50 or more hits (domains). Seems
like a lot to hope that 200 traces will be enough to suss out the
edges of all those hits.  Added long_target (nhmmer) specific code
to increase the number of traces for windows with large nexpected
values (expected # of domains), doing MAX(200, 20*nexpected) traces.

(2) In testing, I found that changes in the size of a window passed to
p7_domaindef_ByPosteriorHeuristics() resulted in wildly different results 
for the null2 calculations coming out of p7_Null2_ByTrace(), such that, 
for one the resulting null2-based domcorrection in rescore_isolated_domain()
was 1.5 bits for one window size and 18.5 bits for another. As this is 
likely also related to the presence of a large number of domains in a
long window, I've modified rescore_isolated_domain() to recompute null2
for each domain using p7_Null2_ByExpectation() for all long_target 
(nhmmer) calls.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3831 db3e70e4-12e1-0310-b23f-81a6e3d09ba7