Added a new variant of the null3 score modifier, p7_null3_windowed_score().
Instead of using the sequence makeup of the full envelope to compute the
bias contribution of each base, this function bases the bias contribution
of a position i on the observed characters in a window around that 
position. This is intended to address situations in Dfam testing in which
a hit gains all it's score from two low-complexity regions separated by
a mostly don't-care region of the model. If the low-complexity regions
have different signatures (e.g. one is GC-rich and the other is AT-rich),
the composition of the whole hit looks to be unbiased, while sub-sections
of the hit are clearly very biased. 

There's a fair bit of tuning to be done regarding the width of window
used to guage regional bias. This commit allows me to perform a first 
round of large scale Dfam tests. 



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3861 db3e70e4-12e1-0310-b23f-81a6e3d09ba7