Added function p7_hmmfile_WriteToString that converts an P7_HMM can write the ASCII version to a string.
At the moment this is in complete isolation and does
not affect any existing methods as suggested by Sean.
It is required for the HMMER website where it is more
convenient not to write files out, as we then write
the string to the hmmpgmd daemon.

Very basic benchmark tests of running mutliple
rounds of hmmbuild (replacing the contents of
p7_hmmfile_WriteASCII with a call to this new function
followed by a single print and freeing the returned 
string) actually turns out to be marginally quicker, 
depsite the double loop through the HMM struc to 
determine the size of the string.

Also, removed a double test on 

hmm->flags & p7H_STATS

in p7_hmmfile_WriteASCII.


Travis also made me go and fix all of the tabs and
formating in the file - said it would be a good
exercise......



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3907 db3e70e4-12e1-0310-b23f-81a6e3d09ba7