Deleted the following sentence from the description of the '--outformat' option in the hmmalign man page: "Default is to autodetect the format of the file". This doesn't make sense for the output file. I'm guessing it was accidentally left in after a copy and paste.
The description of --outformat is still probably out of date, since it
states that only Stockholm or SELEX are available, but I didn't
change this as it wasn't obviously incorrect. My guess is that this
just hasn't been updated since Sean's msafile work, but maybe it's
worth considering requiring Stockholm or SELEX output in hmmalign, so
that PPs are necessarily included? If HMMER does that, I'll do the
same in cmalign.


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3914 db3e70e4-12e1-0310-b23f-81a6e3d09ba7