Removed hmmpgmd2msa from the list of compiled applications. It's addition in a recent commit was unnecessary, as the key role of that code is as a library called by the HMMER website's "manual" jackhmmer.
Produced a function p7_tracealign_getTracesAndStats() that compute 
traces and stats for a collection of sequences relative to a given 
HMM. The returned stats are specifically arrays of:
 - The posterior probability associated with each residue aligned 
   to the core model
 - The relative entropy of the model position to which each matched 
   residue is aligned
 - The bit score of residue vs the model for each matched residue
These stats are to be used in a tool Rob has devised for marking
up protein structures with information about how the sequence
tied to that structure aligns to the relevant HMM.

A portion of this code directly overlaps the functionality of 
hmmalign. In the spirit of code reuse, I pulled a portion of 
hmmalign's main() function code out into a function called 
p7_tracealign_computeTraces() in tracealign.c. 





git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3930 db3e70e4-12e1-0310-b23f-81a6e3d09ba7