Changed prototype of p7_MSVFilter_longtarget(). It now takes an additional flag: <force_ssv>. If TRUE, sse and vmx versions run SSV, not MSV, regardless of score threshold. Previously, and currently with <force_ssv> = FALSE, SSV is run only if the P value equates to a score that reaches a threshold (the score for using the J state).
There's a wrinkle here for the dummy implementation. I had to add
<force_ssv> to the dummy version as well, but no generic SSV exists,
so it has no effect if it is TRUE or FALSE. Currently the dummy
version will never be called with a <do_force> = TRUE value, because
all HMMER calls use <do_force> = FALSE. It's only Infernal calls that
will set it to TRUE, and Infernal doesn't allow dummy builds (or won't
in the near future at least).

The motivation for this change is Infernal-related. For the 1973 Rfam
10.1 models, the corresponding filter HMMs all use SSV instead of MSV
(even before the addition of <force_ssv>) because the F1 P-value
filter threshold is sufficiently high that the score required to pass
SSV does not exceed the score for using the J state. This is true for
any F1 value above 0.00247 for all Rfam 10.1 models. The smallest
default F1 used in Infernal is 0.05, so this means that only users
manually setting F1 very low would ever have used MSV anyway. The
reason I'm forcing SSV here is to be safe; so that MSV is never run
for Infernal. I don't want MSV run because currently it does not set
window diagonal values the same way SSV does -- Travis is still
working to resolve this. So... this change is only really necessary
because the Infernal 1.1 release is occuring while the p7 long target
pipeline is still being modified.

ref: ~nawrockie/notebook/12_0318_inf_final_features/00LOG


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3957 db3e70e4-12e1-0310-b23f-81a6e3d09ba7