This commit removes the 'haircut' code I commited a couple weeks ago. In that code, hmmbuild accepted the flags --hc_start and --hc_end, and masked the corresponding range of positions in the model, where 'mask' means 'make emission probabilities equal to the background'. That approach was a temporary band-aid to get results for the GIRI meeting, but had problems (needed to know the model positions instead of alignment columns; masked position information didn't pass on to downstram alignments; couldn't mask multiple stretches). The new approach uses a GC stockholm entry, akin to the 'RF' line, called 'MM' for 'Model Mask'. A 'MM' line is the length of the alignment, and consists of '.' and 'm' characters; columns corresponding to the 'm' characters are masked - driving all characters in the column to the degenerate character for the alphabet.
This change touches a lot of files, as the MM string needs to be
carried through the msa, hmm, and general/optimized matrix data 
structures in order to be propogated to downstream alignments, 
e.g. in hmmalign. In particular, carrying the MM annotation into
the HMM required producing a new HMM file version (p7_HMMFILE_3f).
The change basically consisted of hunting down every occurence of 
each of the following strings in the code, and either appending 
the appropriate MM-related code, or adding new code for 
validating the new file-format:
 ->rf; v3e_[pf]?magic; p7_HMMFILE_3e; HMMER3/e  

The MM-related changes are primarily self-contained, but one
side-effect is that the consensus string stored in the HMM for
an alignment is converted to the degenerate ambiguity code ('N'
for DNA, 'X' for protein) for every match state corresponding
to a masked ('m') position in the alignment.




git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@3960 db3e70e4-12e1-0310-b23f-81a6e3d09ba7