Fairly meaty change to windowing in the long-target pipeline used by nhmmer. The motivation was that the MSV windowing step generally produces faily short windows, but can produce exceptionally long ones for some of the longer and more-prevelant transposable element models. I've observed windows of length 600Kb making it out the other end of MSV windowing step - that's bad both because it means a huge chunk of sequence makes it to the Forward filter and postprocessing, and also because of numerical stability issues (Sean mentioned problems in Forward at lengths >100Kb)
The idea is a lot like the scanning approach used in 
p7_MSVFilter_longtarget(), and is centered around a new function, 
p7_ViterbiFilter_longtarget(). Basically, the function takes 
an MSV-produced window, determines the score required to pass 
the Viterbi stage (i.e. to meet --F2, 0.1% by default), then
scans the window until it finds a local alignment meeting that
score threshold ... then resets the score and continues the 
scan. Around each such high-scoring position, a window is 
formed. Typically these windows are much shorter, but in the 
case of windows >80Kb, the window is broken into overlapping 
chunks <=40Kb. These shorter windows are then passed on to 
the Forward filter, and possibly post-processing.

(the change to configure.ac is simply an update to the date,
done because I'm releasing an alpha to a few testers, and 
"Nov 2011" was starting to look old)



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4018 db3e70e4-12e1-0310-b23f-81a6e3d09ba7