Changed version from 3.1dev to i1.1rc1, and date to June 2012.
hmmer's configure is not normally run in an build of Infernal, but it
can be if a sneaky user goes into infernal/hmmer/ and manually runs
configure. 

With the change in this revision, HMMER programs that get built
following an infernal/hmmer/configure will have version info set as
i1.1rc1, not 3.1dev. 

(Remember this is not the trunk, but an infernal release branch).


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/infernal@4122 db3e70e4-12e1-0310-b23f-81a6e3d09ba7