Merged changes from Infernal 1.1 branch back to trunk by executing the following command:
svn merge -r 4123:4124 https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/infernal/1.1

Change is to check if documentation/userguide directory exists before
trying to build a Makefile in it. This directory gets removed in a
distro release, so it's important configure doesn't try to create a
Makefile in there (if it does, configure will fail).


git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4134 db3e70e4-12e1-0310-b23f-81a6e3d09ba7