sparse_fwdback_utest test driver added, running _compare_reference() and _singlepath() tests. (xref J10/33).
api for adding cells to a sparse matrix added to p7_sparsemx.[ch]. api
is counterintuitive because production code needs to add cells in
reverse order, because it does posterior decoding in a linear-memory
backwards pass. The SPARSEMASK structure stores its list of cells in
reverse order, so traversing it looks odd.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/sre-banding@4147 db3e70e4-12e1-0310-b23f-81a6e3d09ba7