Query name was missing from the Dfamout output table format; added it.
When run with --cut_ga (and nc and tc), nhmmer was failing to remove 
duplicate hits (the result of searching overlapping windows), because 
their IS_REPORTED and IS_INCLUDED flags were set prior to duplicate 
identification. Modified p7_tophits_RemoveDuplicates() to correctly 
reset those flags for duplicates. 

Also the hit tally in the summary output was failing to account for 
duplicate removal, due to checking the wrong flags. Fixed.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4149 db3e70e4-12e1-0310-b23f-81a6e3d09ba7