Merging changes in r4148 and r4150 (but not r4149) from HMMER's trunk to the HMMER Infernal 1.1 release branch.
I think ideally we'd make these in the HMMER Infernal 1.1 branch
first, and then merge them to the HMMER trunk.  But that doesn't quite
make sense in this case, because Travis is developing on the trunk and
I need to pick and choose the trunk revisions that get merged to the
Infernal 1.1 release branch, so that's why I'm merging from the trunk
to the branch.

r4148 log from hmmer trunk:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Fixed a race condition in the test i20-fmindex-core.pl, where stdout and sterr
were being directed to the same output, which was then being compared to 
expected output from the fmindex exactmatch. The check doesn't care about
the stderr output, just the stdout output, so I've stopped them from mixing. 
Shoving stderr output into the middle of the stdout stream breaks the test.
I have no idea why this test hasn't been intermittently failing for months, 
since the placement of stderr into stdout is the a race condition
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

r4150 log from hmmer trunk:
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
When running nhmmer using a concatenated hmm file (multiple hmms in one file), 
E-value calculation was broken. The calculation depends on keeping a count
of the number of residues searched, which requires that the count be reset
to zero for each new model ... but that wasn't happening. So the E-value for
the second model was twice actual value, the third model was 3x actual 
value, and so forth. Fixed.
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/branches/infernal/1.1@4162 db3e70e4-12e1-0310-b23f-81a6e3d09ba7