The focus of this commit is a new approach for handling composition bias, now accounting for bias as DP progresses, rather than doing it as a post-processing step. This is vital for a large collection of repeats - if we only estimate a bias correction after the alignment is produced, many alignments extend deeply into low-complexity sequence. The final score accounts for this, but the (often atrocious) boundaries have already been set.
In the new approach, just before alignments are produced (in rescore_isolated_domain), we establish new background priors based on the envelope of the target, and change match state emission log-odds scores accordingly. This essentially produces the null3 score correction, but does so at alignment time, instead of after. To overcome the fact that an envelope can sometimes be extremely biased as a result of being short, the envelope's residue frequency is mixed with a prior (either uniform or one based on the long sequence block from with the short envelope has been extracted). A few dozen anecdotal results are very promising, suggesting that this works much better than the previous crummy ways of trying to trim in problematic alignment ends.

This approach only applies to nhmmer. Care was taken to limit overhead cost, and speed impact is negligible. To avoid overwhelming allocations, the bg and om are modified in place. This required creating a true copy of the om for each thread, rather than the old approach of cloning the om.

These changes also involve removal of APP/hq trimming method and trim_bits trimming method, as these are unsatisfactory.



git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@4165 db3e70e4-12e1-0310-b23f-81a6e3d09ba7